+++ app/dashboard/src/components/Filters.tsx
@@ -102,7 +102,7 @@ export const Filters: FC<FilterProps> = ({ ...props }) => {
               })}
             />
           </IconButton>
-          <Button colorScheme="primary" size="sm" onClick={() => onCreateUser(true)} px={5}>
+          <Button bgColor="primary.600" color="gray.100" size="sm" onClick={() => onCreateUser(true)} px={5}>
             {t("createUser")}
           </Button>
         </HStack>
