+++ app/dashboard/src/pages/Users.tsx
@@ -10,18 +10,18 @@ import { QRCodeDialog } from "components/QRCodeDialog";
 import { ResetAllUsageModal } from "components/ResetAllUsageModal";
 import { ResetUserUsageModal } from "components/ResetUserUsageModal";
 import { RevokeSubscriptionModal } from "components/RevokeSubscriptionModal";
-import { Statistics } from "components/Statistics";
 import { UserDialog } from "components/UserDialog";
 import { UsersTable } from "components/UsersTable";
 import { NodesDialog } from "components/nodes";
+import { UserStatistics } from "components/statistics/UserStatistics";
 import { FC } from "react";
 
 export const Dashboard: FC = () => {
   return (
     <VStack justifyContent="space-between" minH="full" rowGap={4}>
       <Box w="full">
-        <Header />
-        <Statistics mt="4" />
+        <Header pageName="users" />
+        <UserStatistics mt="4" />
         <Filters />
         <UsersTable />
         <UserDialog />
